backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "alembic>=1.17.2",
    "datasets>=4.4.1",
    "evaluate>=0.4.6",
    "fastapi-utils>=0.8.0",
    "fastapi[standard]>=0.119.1",
    "kaggle>=1.7.4.5",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers[torch]==4.44.2",
    "typing-inspect>=0.9.0",
]
